backend/requirements.txt
fastapi==0.115.0
python-dotenv==1.0.1
pymongo==4.10.1
python-dateutil==2.8.2
httpx==0.27.2
cloudinary==1.41.0
python-multipart